finished expense form

+++ api/migrations/001_create_my_table.py
@@ -69,8 +69,8 @@ steps = [
             id SERIAL PRIMARY KEY NOT NULL,
             title VARCHAR(200) NOT NULL,
             date DATE NOT NULL,
-            expense_total INT NOT NULL,
-            expense_converted INT NOT NULL,
+            expense_total DECIMAL NOT NULL,
+            expense_converted DECIMAL NOT NULL,
             description TEXT NULL,
             budget_id INT REFERENCES budgets(id),
             category_id INT REFERENCES categories(id)
